Use ParseURL from rueidis package
